requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
discord-handler==0.0.5
flake8==4.0.1
idna==3.3
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
py-cord==2.0.0b7
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
requests==2.27.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
